[syft] move node.credentials to types
