+++ packages/syft/src/syft/client/api.py
@@ -22,8 +22,6 @@ from result import Result
 from typeguard import check_type
 
 # relative
-from ..node.credentials import SyftSigningKey
-from ..node.credentials import SyftVerifyKey
 from ..serde.deserialize import _deserialize
 from ..serde.recursive import index_syft_by_module_name
 from ..serde.serializable import serializable
@@ -38,6 +36,8 @@ from ..service.response import SyftSuccess
 from ..service.service import UserLibConfigRegistry
 from ..service.service import UserServiceConfigRegistry
 from ..types.abstract_node import AbstractNode
+from ..types.credentials import SyftSigningKey
+from ..types.credentials import SyftVerifyKey
 from ..types.syft_object import SYFT_OBJECT_VERSION_1
 from ..types.syft_object import SyftBaseObject
 from ..types.syft_object import SyftObject
