+++ packages/syft/src/syft/external/oblv/oblv_keys_stash.py
@@ -8,7 +8,6 @@ from result import Ok
 from result import Result
 
 # relative
-from ...node.credentials import SyftVerifyKey
 from ...serde.serializable import serializable
 from ...service.response import SyftError
 from ...store.document_store import BaseStash
@@ -16,6 +15,7 @@ from ...store.document_store import DocumentStore
 from ...store.document_store import PartitionSettings
 from ...store.document_store import QueryKeys
 from ...store.document_store import UIDPartitionKey
+from ...types.credentials import SyftVerifyKey
 from ...types.uid import UID
 from .oblv_keys import OblvKeys
