+++ packages/syft/src/syft/node/routes.py
@@ -23,9 +23,9 @@ from ..service.user.user import UserCreate
 from ..service.user.user import UserPrivateKey
 from ..service.user.user_service import UserService
 from ..types.abstract_node import AbstractNode
+from ..types.credentials import SyftVerifyKey
+from ..types.credentials import UserLoginCredentials
 from ..util.telemetry import TRACE_MODE
-from .credentials import SyftVerifyKey
-from .credentials import UserLoginCredentials
 from .worker import Worker
