+++ packages/syft/src/syft/node/worker_settings.py
@@ -5,10 +5,10 @@ from __future__ import annotations
 from typing_extensions import Self
 
 # relative
-from ..node.credentials import SyftSigningKey
 from ..serde.serializable import serializable
 from ..store.document_store import StoreConfig
 from ..types.abstract_node import AbstractNode
+from ..types.credentials import SyftSigningKey
 from ..types.syft_object import SYFT_OBJECT_VERSION_1
 from ..types.syft_object import SyftObject
 from ..types.uid import UID
