+++ packages/syft/src/syft/service/action/action_permissions.py
@@ -3,8 +3,8 @@ from enum import Enum
 from typing import Optional
 
 # relative
-from ...node.credentials import SyftVerifyKey
 from ...serde.serializable import serializable
+from ...types.credentials import SyftVerifyKey
 from ...types.uid import UID
