+++ packages/syft/src/syft/service/data_subject/data_subject_member_service.py
@@ -7,12 +7,12 @@ from typing import Union
 from result import Result
 
 # relative
-from ...node.credentials import SyftVerifyKey
 from ...serde.serializable import serializable
 from ...store.document_store import BaseUIDStoreStash
 from ...store.document_store import DocumentStore
 from ...store.document_store import PartitionSettings
 from ...store.document_store import QueryKeys
+from ...types.credentials import SyftVerifyKey
 from ...util.telemetry import instrument
 from ..context import AuthedServiceContext
 from ..response import SyftError
