+++ packages/syft/src/syft/service/dataset/dataset_stash.py
@@ -6,13 +6,13 @@ from typing import Optional
 from result import Result
 
 # relative
-from ...node.credentials import SyftVerifyKey
 from ...serde.serializable import serializable
 from ...store.document_store import BaseUIDStoreStash
 from ...store.document_store import DocumentStore
 from ...store.document_store import PartitionKey
 from ...store.document_store import PartitionSettings
 from ...store.document_store import QueryKeys
+from ...types.credentials import SyftVerifyKey
 from ...types.uid import UID
 from ...util.telemetry import instrument
 from .dataset import Dataset
