+++ packages/syft/src/syft/service/message/messages.py
@@ -4,9 +4,9 @@ from typing import Optional
 
 # relative
 from ...client.api import APIRegistry
-from ...node.credentials import SyftVerifyKey
 from ...serde.serializable import serializable
 from ...store.linked_obj import LinkedObject
+from ...types.credentials import SyftVerifyKey
 from ...types.datetime import DateTime
 from ...types.syft_object import SYFT_OBJECT_VERSION_1
 from ...types.syft_object import SyftObject
