+++ packages/syft/src/syft/service/metadata/node_metadata.py
@@ -11,8 +11,8 @@ from packaging import version
 from pydantic import BaseModel
 
 # relative
-from ...node.credentials import SyftVerifyKey
 from ...serde.serializable import serializable
+from ...types.credentials import SyftVerifyKey
 from ...types.syft_object import SYFT_OBJECT_VERSION_1
 from ...types.syft_object import StorableObjectType
 from ...types.syft_object import SyftObject
