+++ packages/syft/src/syft/service/policy/user_policy_stash.py
@@ -5,12 +5,12 @@ from typing import List
 from result import Result
 
 # relative
-from ...node.credentials import SyftVerifyKey
 from ...serde.serializable import serializable
 from ...store.document_store import BaseUIDStoreStash
 from ...store.document_store import DocumentStore
 from ...store.document_store import PartitionSettings
 from ...store.document_store import QueryKeys
+from ...types.credentials import SyftVerifyKey
 from .policy import PolicyUserVerifyKeyPartitionKey
 from .policy import UserPolicy
