+++ packages/syft/src/syft/service/project/project.py
@@ -25,10 +25,10 @@ from typing_extensions import Self
 # relative
 from ...client.client import SyftClient
 from ...client.client import SyftClientSessionCache
-from ...node.credentials import SyftSigningKey
-from ...node.credentials import SyftVerifyKey
 from ...serde.serializable import serializable
 from ...service.metadata.node_metadata import NodeMetadata
+from ...types.credentials import SyftSigningKey
+from ...types.credentials import SyftVerifyKey
 from ...types.datetime import DateTime
 from ...types.syft_object import SYFT_OBJECT_VERSION_1
 from ...types.syft_object import SyftObject
