+++ packages/syft/src/syft/service/project/project_stash.py
@@ -6,13 +6,13 @@ from typing import Optional
 from result import Result
 
 # relative
-from ...node.credentials import SyftVerifyKey
 from ...serde.serializable import serializable
 from ...store.document_store import BaseUIDStoreStash
 from ...store.document_store import PartitionKey
 from ...store.document_store import PartitionSettings
 from ...store.document_store import QueryKeys
 from ...store.document_store import UIDPartitionKey
+from ...types.credentials import SyftVerifyKey
 from ...types.uid import UID
 from ...util.telemetry import instrument
 from ..request.request import Request
