+++ packages/syft/src/syft/service/request/request.py
@@ -18,10 +18,10 @@ from typing_extensions import Self
 # relative
 from ...client.api import APIRegistry
 from ...external import OBLV
-from ...node.credentials import SyftVerifyKey
 from ...serde.serializable import serializable
 from ...serde.serialize import _serialize
 from ...store.linked_obj import LinkedObject
+from ...types.credentials import SyftVerifyKey
 from ...types.datetime import DateTime
 from ...types.syft_object import SYFT_OBJECT_VERSION_1
 from ...types.syft_object import SyftObject
