+++ packages/syft/src/syft/service/user/user.py
@@ -12,9 +12,9 @@ import pydantic
 from pydantic.networks import EmailStr
 
 # relative
-from ...node.credentials import SyftSigningKey
-from ...node.credentials import SyftVerifyKey
 from ...serde.serializable import serializable
+from ...types.credentials import SyftSigningKey
+from ...types.credentials import SyftVerifyKey
 from ...types.syft_object import PartialSyftObject
 from ...types.syft_object import SYFT_OBJECT_VERSION_1
 from ...types.syft_object import SyftObject
