+++ packages/syft/src/syft/store/kv_document_store.py
@@ -16,7 +16,6 @@ from result import Result
 from typing_extensions import Self
 
 # relative
-from ..node.credentials import SyftVerifyKey
 from ..serde.serializable import serializable
 from ..service.action.action_permissions import ActionObjectEXECUTE
 from ..service.action.action_permissions import ActionObjectOWNER
@@ -25,6 +24,7 @@ from ..service.action.action_permissions import ActionObjectREAD
 from ..service.action.action_permissions import ActionObjectWRITE
 from ..service.action.action_permissions import ActionPermission
 from ..service.response import SyftSuccess
+from ..types.credentials import SyftVerifyKey
 from ..types.syft_object import SyftObject
 from ..types.uid import UID
 from .document_store import BaseStash
