+++ packages/syft/tests/syft/messages/fixtures.py
@@ -2,8 +2,6 @@
 import pytest
 
 # syft absolute
-from syft.node.credentials import SyftSigningKey
-from syft.node.credentials import SyftVerifyKey
 from syft.node.worker import Worker
 from syft.service.context import AuthedServiceContext
 from syft.service.message.message_service import MessageService
@@ -12,6 +10,8 @@ from syft.service.message.messages import CreateMessage
 from syft.service.message.messages import Message
 from syft.service.user.user import User
 from syft.store.linked_obj import LinkedObject
+from syft.types.credentials import SyftSigningKey
+from syft.types.credentials import SyftVerifyKey
 from syft.types.datetime import DateTime
 from syft.types.uid import UID
