+++ packages/syft/tests/syft/request/fixtures.py
@@ -4,11 +4,11 @@ import pytest
 
 # syft absolute
 from syft.client.client import SyftClient
-from syft.node.credentials import SyftVerifyKey
 from syft.node.worker import Worker
 from syft.service.context import AuthedServiceContext
 from syft.service.request.request_stash import RequestStash
 from syft.store.document_store import DocumentStore
+from syft.types.credentials import SyftVerifyKey
 
 
 @pytest.fixture
