+++ packages/syft/tests/syft/settings/settings_service_test.py
@@ -7,13 +7,13 @@ from result import Err
 from result import Ok
 
 # syft absolute
-from syft.node.credentials import SyftVerifyKey
 from syft.service.context import AuthedServiceContext
 from syft.service.response import SyftError
 from syft.service.settings.settings import NodeSettings
 from syft.service.settings.settings import NodeSettingsUpdate
 from syft.service.settings.settings_service import SettingsService
 from syft.service.settings.settings_stash import SettingsStash
+from syft.types.credentials import SyftVerifyKey
 
 
 def test_settingsservice_get_success(
