+++ packages/syft/tests/syft/users/user_stash_test.py
@@ -2,12 +2,12 @@
 from faker import Faker
 
 # syft absolute
-from syft.node.credentials import SyftSigningKey
 from syft.service.response import SyftSuccess
 from syft.service.user.user import User
 from syft.service.user.user import UserUpdate
 from syft.service.user.user_roles import ServiceRole
 from syft.service.user.user_stash import UserStash
+from syft.types.credentials import SyftSigningKey
 from syft.types.uid import UID
